fix: do not erase systemd environment variables
In some scenarios, one might need to call ListenFd::from_env many times. Due to
those two problematic lines, this is impossible, as the first time will erase
the environment, and all the others will be ListenFd::empty.

In case where one wants to call this from a library, or any such thing, it
makes the library completely unusable. There is also no reason for it to mutate
the environment like that. Keeping those variables has effectively no impact.